Refactored Authentication code
- No need to manually store key in AuthSession after calling Authenticate()
